Lots of timer stuff.
Fixed a "feature" in pc_system.cc with setting timers to small values
that can cause bochs to hang.

Significantly improved the performance of the new PIT.
It's probably ready to become the default now.

Added a preliminary implementation of the slowdown timer
that Bryce and I had talked about.


git-svn-id: svn://svn.code.sf.net/p/bochs/code/trunk@1169 6c8e4198-a580-4521-a02f-c9ae86df1db9