Latest round of patches/fixups to get 64 bit emulation further.
This is an interim update to allow others to test.

We have userland code running!!! (up to a point)

Able to start executing "sash" as /sbin/init in userland from linux 64 bit
kernel until it crashes trying to access a null pointer.  No kernel panics
though, just a segfault loop.


git-svn-id: svn://svn.code.sf.net/p/bochs/code/trunk@3413 6c8e4198-a580-4521-a02f-c9ae86df1db9